Refactor project structure to use build directory for Electron and renderer; update paths and configurations
