+++ index.html
@@ -9,7 +9,7 @@
 
 <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="src/renderer/main.tsx"></script>
 </body>
 
 </html>
