+++ src/renderer/App.tsx
@@ -1,12 +1,12 @@
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom'
-import ChooseProject from './components/select_project/SelectProject'
-import CreateProject from './components/select_project/CreateProject'
-import LoadProject from './components/select_project/LoadProject'
-import ImportProject from './components/select_project/ImportProject'
-import Inspection from './components/Inspection/Inspection'
-import Report from './components/Inspection/Report'
+import ChooseProject from './pages/SelectProject'
+import CreateProject from './pages/CreateProject'
+import LoadProject from './pages/LoadProject'
+import ImportProject from './pages/ImportProject'
+import Inspection from './pages/Inspection'
+import Report from './pages/Report'
 import AppHeader from './components/AppHeader'
-import Home from './components/Home'
+import Home from './pages/Home'
 import AppFooter from './components/AppFooter'
 
 const App = () => {
