+++ src/renderer/assets/css/Inspection/Inspection.css
