+++ src/renderer/components/AppFooter.tsx
