+++ src/renderer/components/AppHeader.tsx
