+++ tsconfig.json
@@ -24,10 +24,10 @@
     "jsx": "react-jsx"
   },
   "include": [
-    "./src"
+    "src"
   ],
   "exclude": [
-    "./electron"
+    "src/main"
   ],
   "references": [
     {
